# swagger.yaml
openapi: 3.0.3
info:
  title: Alliance Courtage API
  description: |
    API REST pour l'extranet Alliance Courtage.
    
    Cette API permet de gérer :
    - L'authentification et la gestion des utilisateurs
    - Les produits financiers et structurés
    - Les documents et archives
    - Les formations et notifications
    - Les simulateurs fiscaux
    - Le contenu CMS
    - Les réservations de produits
    
    **Authentification** : La plupart des endpoints nécessitent un token JWT dans le header `x-auth-token`.
  version: 1.0.0
  contact:
    name: Alliance Courtage
    email: support@alliance-courtage.fr
servers:
  - url: http://localhost:3001/api
    description: Serveur de développement local
  - url: https://api.alliance-courtage.fr/api
    description: Serveur de production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT dans le header x-auth-token
  schemas:
...
    SimulatorUsage:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        simulator_type:
          type: string
          enum: [ir, ifi, succession, placement]
        parameters:
          type: object
          nullable: true
        result_summary:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
paths:
...
  /simulators/usage:
    post:
      tags:
        - Simulators
      summary: Enregistrer l'accès à un simulateur
      description: Enregistrer qu'un utilisateur a accédé à un simulateur
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - simulator_type
              properties:
                simulator_type:
                  type: string
                  enum: [ir, ifi, succession, placement]
                  description: Type de simulateur
                parameters:
                  type: object
                  nullable: true
                  description: Paramètres utilisés (optionnel)
                result_summary:
                  type: string
                  nullable: true
                  description: Résumé du résultat (optionnel)
      responses:
        '200':
          description: Accès enregistré avec succès
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Utilisation du simulateur enregistrée avec succès
        '400':
          description: Données invalides
        '401':
          description: Non authentifié

    get:
      tags:
        - Simulators
      summary: Liste des utilisations de simulateurs
      description: Obtenir la liste détaillée des utilisations de simulateurs (Admin uniquement)
      security:
        - bearerAuth: []
      parameters:
        - name: simulator_type
          in: query
          schema:
            type: string
            enum: [ir, ifi, succession, placement]
        - name: user_id
          in: query
          schema:
            type: integer
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        '200':
          description: Liste des utilisations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SimulatorUsage'
        '401':
          description: Non authentifié
        '403':
          description: Accès refusé (admin uniquement)
  /simulators/usage/stats:
    get:
      tags:
        - Simulators
      summary: Statistiques agrégées des simulateurs
      description: Obtenir les statistiques agrégées d'utilisation des simulateurs (Admin uniquement)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Statistiques agrégées
          content:
            application/json:
              schema:
                type: object
                properties:
                  by_type:
                    type: array
                    items:
                      type: object
                      properties:
                        simulator_type:
                          type: string
                        total_uses:
                          type: integer
                        unique_users:
                          type: integer
                        first_use:
                          type: string
                          format: date-time
                        last_use:
                          type: string
                          format: date-time
                  by_user:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        nom:
                          type: string
                        prenom:
                          type: string
                        email:
                          type: string
                        total_uses:
                          type: integer
                        simulators_used:
                          type: integer
                        last_use:
                          type: string
                          format: date-time
                  global:
                    type: object
                    properties:
                      total_uses:
                        type: integer
                      total_users:
                        type: integer
                      total_simulators:
                        type: integer
                      first_use_ever:
                        type: string
                        format: date-time
                      last_use_ever:
                        type: string
                        format: date-time
                  daily:
                    type: array
                    items:
                      type: object
                      properties:
                        date:
                          type: string
                          format: date
                        uses:
                          type: integer
                        users:
                          type: integer
        '401':
          description: Non authentifié
        '403':
          description: Accès refusé (admin uniquement)
...